+++ src/scripts/main-content.ts
@@ -4,3 +4,4 @@ import '../styles/main-content.scss';
 // scripts
 import './border';
 import './overlay';
+import './measurement';